+++ src/main/java/net/dreamlu/mica/auto/common/BootAutoType.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright (c) 2019-2029, Dreamlu 卢春梦 (596392912@qq.com & www.dreamlu.net).
+ * <p>
+ * Licensed under the GNU LESSER GENERAL PUBLIC LICENSE 3.0;
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ * http://www.gnu.org/licenses/lgpl.html
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.dreamlu.mica.auto.common;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import net.dreamlu.mica.auto.annotation.AutoContextInitializer;
+import net.dreamlu.mica.auto.annotation.AutoFailureAnalyzer;
+import net.dreamlu.mica.auto.annotation.AutoListener;
+
+/**
+ * 注解类型
+ *
+ * @author L.cm
+ */
+@Getter
+@AllArgsConstructor
+public enum BootAutoType {
+	/**
+	 * 注解处理的类型
+	 */
+	CONTEXT_INITIALIZER(AutoContextInitializer.class.getName(), "org.springframework.context.ApplicationContextInitializer"),
+	LISTENER(AutoListener.class.getName(), "org.springframework.context.ApplicationListener"),
+	FAILURE_ANALYZER(AutoFailureAnalyzer.class.getName(), "org.springframework.boot.diagnostics.FailureAnalyzer"),
+	COMPONENT("org.springframework.stereotype.Component", "org.springframework.boot.autoconfigure.EnableAutoConfiguration");
+
+	private final String annotation;
+	private final String configureKey;
+
+}